backend/requirements.txt
Flask
flask-cors
python-dotenv
pymongo
pyjwt
Flask-CORS
Flask-Limiter
flask-login
